requirements.txt
openai>=1.30.0
pyaudio>=0.2.14
pyinstaller>=6.0
pynput>=1.7.6
pyperclip>=1.8.2
python-dotenv>=1.0.0
pyyaml>=6.0
rumps>=0.4.0
